Switch scaling factors back to floats, as they should be.

+++ Trinity.Encore.Framework.Game/IO/Formats/DBC/SpellScalingRecord.cs
@@ -40,14 +40,13 @@ namespace Trinity.Encore.Framework.Game.IO.Formats.DBC
             return castTime;
         }
 
-        // shouldn't these be floats?
         public sealed class SpellScalingData
         {
-            public int Part1 { get; set; }
+            public float Part1 { get; set; }
 
-            public int Part2 { get; set; }
+            public float Part2 { get; set; }
 
-            public int Part3 { get; set; }
+            public float Part3 { get; set; }
         }
     }
 }
